npm audit fix --force
this may also fix a unidici type error "terminated" thing I'm getting in who's that pokemon intial printing:  [cause]: TypeError [ERR_INVALID_ARG_TYPE]: The "stream" argument must be an instance of Stream. Received an instance of ReadableStream

I say that just because the audit is upgrading unidici, so that might incidentally get fixed, not because I know anything about the problem.