Partial (and "evil")-revert v3.7.0 "attach, per se, discord attachment links, in send_long", which did not, in fact, work.
The problem here is that it was fetching the content from discord at the bare url and then sending that to the user. This was thus the 36-byte message "This content is no longer available." Which is not desireable!

It is extremely difficult to figure out Discord and DiscordJS's wacky-tabacky byzantine API, to figure out how to do this the right way, if that's even possible. So, instead, we will rely on auto-embedding and simply give up the feature of having auto show up in the audio player (which, honestly, should be a discord feature request anyway).

This reverts commit c259e8eba8a6af76d6923788bf5c760c87e4d91e.